+++ Wrappers/Python/test/test_DataProcessor.py
@@ -3588,15 +3588,15 @@ class TestFluxNormaliser(unittest.TestCase):
             .set_panel([3,3])
         arr = numpy.array([[[1,2,3],[1,2,3],[1,2,3]],
                         [[4,5,6],[4,5,6],[4,5,6]], 
-                        [[7,8,9],[7,8,9],[7,8,9]]])
+                        [[7,8,9],[7,8,9],[7,8,9]]],dtype=numpy.float32)
         self.data_simple = AcquisitionData(arr, geometry=ag)
 
         source_position_set=[[0,-100000,0]]*3
         detector_position_set=[[0,0,0]]*3
         detector_direction_x_set=[[1, 0, 0]]*3
         detector_direction_y_set=[[0, 0, 1]]*3
-        cone_flex_ag = AcquisitionGeometry.create_Cone3D_Flex(source_position_set, detector_position_set, detector_direction_x_set, detector_direction_y_set).set_panel([3,3])
-        self.cone_flex  = AcquisitionData(arr, geometry=cone_flex_ag)
+        cone_flex_ag = AcquisitionGeometry.create_Cone3D_Flex(source_position_set, detector_position_set, detector_direction_x_set, detector_direction_y_set).set_panel([128,128])
+        self.cone_flex  = cone_flex_ag.allocate(1)
 
     def error_message(self,processor, test_parameter):
             return "Failed with processor " + str(processor) + " on test parameter " + test_parameter
@@ -3622,11 +3622,6 @@ class TestFluxNormaliser(unittest.TestCase):
         processor = FluxNormaliser()
         with self.assertRaises(ValueError):
             processor.check_input(self.data_cone)
-
-        # check there's a not implemented error if cone flex geom is used:
-        processor = FluxNormaliser(flux=[1,2,3])
-        with self.assertRaises(NotImplementedError):
-            processor.check_input(self.cone_flex)
 
     def test_calculate_flux(self):
         # check there is an error if flux array size is not equal to the number of angles in data
@@ -3814,7 +3809,8 @@ class TestFluxNormaliser(unittest.TestCase):
 
         # Test no error with preview_configuration with different data shapes
         for data in [self.data_cone, self.data_parallel, self.data_multichannel, 
-                     self.data_slice, self.data_reorder, self.data_single_angle]:
+                     self.data_slice, self.data_reorder, self.data_single_angle,
+                     self.cone_flex]:
             mock_show.reset_mock()
 
             roi = {'horizontal':(25,40)}
@@ -3826,15 +3822,15 @@ class TestFluxNormaliser(unittest.TestCase):
 
             # for 3D, check no error specifying a single angle to plot
             if data.geometry.dimension == '3D':
-                processor.preview_configuration(angle=1)
+                processor.preview_configuration(projection_index=1)
             # if 2D, attempt to plot single angle should cause error
             else:
                 with self.assertRaises(ValueError):
-                    processor.preview_configuration(angle=1)
+                    processor.preview_configuration(projection_index=1)
 
             # if data is multichannel, check no error specifying a single channel to plot
             if 'channel' in data.dimension_labels:
-                processor.preview_configuration(angle=1, channel=1)
+                processor.preview_configuration(projection_index=1, channel=1)
                 processor.preview_configuration(channel=1)
             # if single channel, check specifying channel causes an error
             else:
@@ -3842,6 +3838,25 @@ class TestFluxNormaliser(unittest.TestCase):
                     processor.preview_configuration(channel=1)
 
         # Re-enable logging
+        logging.disable(logging.NOTSET)
+
+    @unittest.skipIf(not has_matplotlib, "matplotlib not installed")
+    @patch('matplotlib.pyplot.show')
+    def test_preview_configuration_deprecated_angle(self, mock_show):
+        logging.disable(logging.CRITICAL)
+
+        roi = {'horizontal':(25,40)}
+        processor = FluxNormaliser(roi=roi)
+        processor.set_input(self.data_cone)
+
+        with self.assertWarns(DeprecationWarning):
+            processor.preview_configuration(angle=1)
+
+        processor.preview_configuration(projection_index=1)
+
+        with self.assertRaises(TypeError):
+            processor.preview_configuration(projection_index=1, angle=1)
+
         logging.disable(logging.NOTSET)
 
     def test_FluxNormaliser(self, accelerated=False):
@@ -3855,6 +3870,12 @@ class TestFluxNormaliser(unittest.TestCase):
         data_norm = processor.get_output()
         numpy.testing.assert_allclose(data_norm.array, self.data_cone.array)
         
+        #Test flux with no target on ConeFlex
+        processor = FluxNormaliser(flux=1, accelerated=accelerated)
+        processor.set_input(self.cone_flex)
+        data_norm = processor.get_output()
+        numpy.testing.assert_allclose(data_norm.array, self.cone_flex.array)
+
         #Test flux with target
         processor = FluxNormaliser(flux=10, target=5.0, accelerated=accelerated)
         processor.set_input(self.data_cone)